Update 'createLink' and 'destroyLink' to allow OAuth usage.

+++ src/Functions/destroyLink.js
@@ -1,6 +1,9 @@
 import * as Express from 'express';
+import { info } from 'heroku-logger';
 
 import Link from '../Models/Link';
+import { user, context } from '../helpers';
+import { UnauthorizedError } from 'express-jwt';
 import NotFoundException from '../Exceptions/NotFoundException';
 
 /**
@@ -17,7 +20,15 @@ export default async function destroyLink(req, res) {
     throw new NotFoundException('Not found.');
   }
 
+  if (user(req) && !user(req).hasRole('staff', 'admin')) {
+    throw new UnauthorizedError('credentials_required', {
+      message: `Requires 'staff' or 'admin' role.`,
+    });
+  }
+
   await link.delete();
+
+  info('Link deleted.', { key: link.key, ...context(req) });
 
   return res.status(200).json({ message: 'Link deleted.' });
 }
